+++ packages/core/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the Zowe core SDK package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Added a `requestQueue` queue object to the `ZosmfRestClient` class. [#2543](https://github.com/zowe/zowe-cli/pull/2543)
+- Enhancement: Added the `ZosmfRestClient.setThrottlingOptions` function to allow extenders to configure REST request throttling settings. [#2543](https://github.com/zowe/zowe-cli/pull/2543)
+
 ## `8.27.4`
 
 - BugFix: Updated minimum supported version of Node from 18 to 20. Added Node 24 support. [#2616](https://github.com/zowe/zowe-cli/pull/2616)